feat: add classifier results and match data to results_dic
Adds a function that iterates through the results_dic and runs
each image through the image classifier. It takes the results
from the classifier, cleans up the string and compares it to
see if the intial label is in that results string. Finally,
the classifier results string and the match boolean (as a 0/1)
is added to the dictionary value list for that image.