feat: check if pet label and classifier label are dogs
This uses an input file with dog names to create a dictionary that has
dog names as keys and arbitrary values of 1. We then take the pet_dict
and check if the pet labels match any of the keys in the dog_dict. If
they match, we update the value for that entry to have a 1 in the 3 index.
Otherwise we input a 0 there.  We then run the same test against the
classifier labels in the pet dictionary to fill out the 4th index. The
final result is a pet_dict that looks like the following:
{
  "collie.png": ["collie", "border collie, collie", 1, 1, 1],
  "lynx.png": ["lynx", "cat", 0, 0, 0]
}